# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.bulkIDsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  controllers.bulkToggleRequest:
    properties:
      enabled:
        type: boolean
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  controllers.taskRequest:
    properties:
      command:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      httpBody:
        type: string
      httpHeaders:
        type: string
      httpMethod:
        type: string
      name:
        type: string
      spec:
        type: string
      timeout:
        type: integer
      type:
        enum:
        - HTTP
        - SHELL
        - FUNC
        type: string
    required:
    - command
    - name
    - spec
    - type
    type: object
...
  models.PageData:
    properties:
      pageData: {}
      total:
        type: integer
    type: object
...
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      originUrl:
        type: string
    type: object
...
  models.Task:
    properties:
      command:
        description: URL / shell 命令 / 内置函数名
        type: string
      createTime:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      httpBody:
        type: string
      httpHeaders:
        description: JSON 字符串：{"K":"V"}
        type: string
      httpMethod:
        type: string
      id:
        type: integer
      lastRunAt:
        type: string
      name:
        type: string
      nextRunAt:
        type: string
      spec:
        description: 标准 5 段 cron 表达式
        type: string
      timeout:
        description: 秒
        type: integer
      type:
        description: HTTP / SHELL / FUNC
        type: string
      updateTime:
        type: string
    type: object
  models.TaskLog:
    properties:
      duration:
        description: 毫秒
        type: integer
      endTime:
        type: string
      error:
        type: string
      id:
        type: integer
      output:
        type: string
      startTime:
        type: string
      status:
        description: SUCCESS / FAILED / TIMEOUT
        type: string
      steps:
        description: 'JSON: [{time, level, message}]'
        type: string
      taskId:
        type: integer
      taskName:
        type: string
      trigger:
        description: CRON / MANUAL
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: kite-admin 后台管理系统 API 文档
  title: Kite Admin API
  version: "1.0"
paths:
...
  /task:
    post:
      consumes:
      - application/json
      description: 创建新的定时任务
      parameters:
      - description: 任务信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.taskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 创建定时任务
      tags:
      - 定时任务
  /task/{id}:
    delete:
      description: 删除指定定时任务
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 删除定时任务
      tags:
      - 定时任务
    patch:
      consumes:
      - application/json
      description: 更新指定定时任务的信息
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      - description: 任务信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.taskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 更新定时任务
      tags:
      - 定时任务
  /task/{id}/run:
    post:
      description: 立即手动触发一次任务执行
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: 执行失败
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 手动执行任务
      tags:
      - 定时任务
  /task/{id}/toggle:
    patch:
      description: 切换指定任务的启用/停用状态
      parameters:
      - description: 任务ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
        "404":
          description: 任务不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 切换任务启用状态
      tags:
      - 定时任务
  /task/bulk/delete:
    post:
      consumes:
      - application/json
      description: 批量删除指定的定时任务
      parameters:
      - description: 任务ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.bulkIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除数量
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 批量删除任务
      tags:
      - 定时任务
  /task/bulk/toggle:
    post:
      consumes:
      - application/json
      description: 批量启用或停用定时任务
      parameters:
      - description: 任务ID列表和目标状态
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.bulkToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 影响数量
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 批量切换启用状态
      tags:
      - 定时任务
  /task/funcs:
    get:
      description: 列出所有可用的内置函数（FUNC类型任务使用）
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取内置函数列表
      tags:
      - 定时任务
  /task/log/page:
    get:
      description: 分页查询任务执行日志，支持按任务ID、状态、触发方式筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNo
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 任务ID
        in: query
        name: taskId
        type: string
      - description: 状态（SUCCESS/FAILED/TIMEOUT）
        in: query
        name: status
        type: string
      - description: 触发方式（CRON/MANUAL）
        in: query
        name: trigger
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/models.PageData'
                  - properties:
                      pageData:
                        items:
                          $ref: '#/definitions/models.TaskLog'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 查询任务日志
      tags:
      - 定时任务
  /task/page:
    get:
      description: 分页查询定时任务列表，支持按名称和类型筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNo
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 任务名称（模糊搜索）
        in: query
        name: name
        type: string
      - description: 任务类型（HTTP/SHELL/FUNC）
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/models.PageData'
                  - properties:
                      pageData:
                        items:
                          $ref: '#/definitions/models.Task'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询任务
      tags:
      - 定时任务
  /task/preview-next:
    get:
      description: 返回cron表达式接下来n次的执行时间
      parameters:
      - description: Cron表达式
        in: query
        name: spec
        required: true
        type: string
      - default: 5
        description: 预览次数
        in: query
        name: "n"
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: 表达式无效
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 预览Cron表达式
      tags:
      - 定时任务
  /task/stats:
    get:
      description: 获取任务总数、启用数、今日执行统计和最近执行记录
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 获取任务统计
      tags:
      - 定时任务
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"